adds routes folder, refactors server to router
