+++ package.json
@@ -12,7 +12,7 @@
     "express": "^4.16.4",
     "helmet": "^3.14.0",
     "knex": "^0.15.2",
-    "sqlite3": "^4.0.2"
+    "sqlite3": "^4.0.6"
   },
   "devDependencies": {
     "nodemon": "^1.18.4"
